task-management-backend/requirements.txt
flask
flask-sqlalchemy
flask-marshmallow
flask-migrate
flask-cors
flask-jwt-extended
marshmallow-sqlalchemy
werkzeug